feat: change linting rules to make `console.foo` an error
Possibly contentious change.

Since we have the `debug` module, we generally don't need to use
`console.log` and friends. It's useful during local debugging
but it's also easy to forget to remove those lines so we could
upgrade this should be an error to help ourselves out.

If you find that do need to use it (e.g. for cli output), you can
easily disable the rule for that line with:

```javascript
console.info('Greetings Professor Falken') // eslint-disable-line no-console
```